Create tictactoe.py
#Shortest Mode
You get the moves of a Tic-Tac-Toe Game. Its boring to look at these numbers. The challenge is to print the game.
Input
N Number of moves
M The move. A move shows the field that the player choice. The fields a numberd like this:
0|1|2
3|4|5
6|7|8
First move is always x, then o and so on.
Output
The Output should be a Tic-Tac-Toe Field. Each cell is either a 'x', 'o' or a whitespace. Cells are seperated with a '|'.
Constraints
All games and moves are valid Tic-Tac-Toe Games
Example
Input
9
4
0
1
7
3
5
6
2
8
Output
o|x|o
x|x|o
x|o|x